register user

+++ src/main/java/com/example/AIRecognize/authentication/verify/service/VerifyService.java
@@ -1,5 +1,6 @@
 package com.example.AIRecognize.authentication.verify.service;
 
+import com.example.AIRecognize.authentication.user.repo.UserRepo;
 import com.example.AIRecognize.authentication.verify.dto.VerifyDTO;
 import com.example.AIRecognize.authentication.verify.entity.Verify;
 import com.example.AIRecognize.authentication.verify.repo.VerifyRepo;
@@ -12,14 +13,20 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.security.SecureRandom;
+import java.time.Instant;
+import java.util.Date;
 
 @Service
 @RequiredArgsConstructor
 public class VerifyService {
     private final VerifyRepo verifyRepo;
+    private final UserRepo userRepo;
     private final MailListenerEvent mailListenerEvent;
     @Transactional
     public VerifyDTO sendMail(String email){
+        if (userRepo.existsByEmail(email))
+            throw new CustomException(HttpStatus.BAD_REQUEST, "This email is already registered!");
+
         if (verifyRepo.existsById(email))
             verifyRepo.deleteById(email);
 
@@ -32,5 +39,19 @@ public class VerifyService {
         verifyRepo.save(verify);
 
         return VerifyDTO.dto(verify);
+    }
+
+    public VerifyDTO verifyEmail(VerifyDTO request){
+        String email = request.getMailPhone();
+        Verify verify = verifyRepo.findById(email).orElseThrow(
+                ()-> new CustomException(HttpStatus.BAD_REQUEST, "No exist email!")
+        );
+
+        if (verify.getExpiration().before(Date.from(Instant.now())))
+            throw new CustomException(HttpStatus.BAD_REQUEST, "Your verified time is expired!");
+        else if (!verify.getVerifyCode().equals(request.getVerifyCode()))
+            throw new CustomException(HttpStatus.BAD_REQUEST, "Your verified code is wrong!");
+        verify.setSuccess(true);
+        return VerifyDTO.dto(verifyRepo.save(verify));
     }
 }
